Initial version of Lucene indexing code
Indexer.java walks a directory, parses all the XML files, and produces a Lucene
index in the current directory.

Searcher.java provides a (very) simple command-line interface to search the
index: it just reads queries off of standard in and returns a list of (score,
title) pairs.

I'm using a Makefile here for simplicity (and because I don't know ant well
enough to write a build.xml from scratch).